Split status batch and moderation actions (#37970)
