+++ app/models/admin/account_action.rb
@@ -66,20 +66,6 @@ class Admin::AccountAction < Admin::BaseAction
     end
   end
 
-  def process_strike!
-    @warning = target_account.strikes.create!(
-      account: current_account,
-      report: report,
-      action: type,
-      text: text_for_warning,
-      status_ids: status_ids
-    )
-
-    # A log entry is only interesting if the warning contains
-    # custom text from someone. Otherwise it's just noise.
-    log_action(:create, @warning) if @warning.text.present? && type == 'none'
-  end
-
   def process_reports!
     # If we're doing "mark as resolved" on a single report,
     # then we want to keep other reports open in case they
@@ -129,17 +115,6 @@ class Admin::AccountAction < Admin::BaseAction
 
   def process_queue!
     queue_suspension_worker! if type == 'suspend'
-  end
-
-  def process_notification!
-    return unless warnable?
-
-    UserMailer.warning(target_account.user, warning).deliver_later!
-    LocalNotificationWorker.perform_async(target_account.id, warning.id, 'AccountWarning', 'moderation_warning')
-  end
-
-  def warnable?
-    send_email_notification? && target_account.local?
   end
 
   def status_ids
